fix: signup and login form close when not intended
This fixes a bug in signup and login form. If you select the text inside
an input field by clicking and dragging until the end of the text, and
then you end dragging of the form, the modal is closed becuase it
considers you have clicked outside of the form.

I've added an explicit "close" button.