test: remove test torrents after running the upload torrent test
The file was being removed from the fixtures folder but not the torrent
in the database.

We are trying to clean the E2E env state after running tests. There are
still some things pending to remove. For example, we create a new user
for the test and the newly generated user is not removed.