fix: [#247] search bar not working when in /torrents endpoint
Using the search bar in the /torrents endpoint, now actually calls the API to fetch the torrents
with the search query and displays the result, making the torrents component render again with
the results matching the search prompt.

A key was added to the <NuxtPage> component so the watcher listening for a new search prompt gets triggered. Also,
minor refactoring was applied to the code to follow  good practices and clean up unused imports and variables.