refactor: [#297] Creation date now display as UTC
If the torrent creation date field is present in the torrent file, it is now shown as a UTC date
in a human readable format.

If there is no creation date, a no creation date provided message is displayed.

And if there is an error parsing the creation date or if it is a no valid date, an invalid date
text is shown.