ci: [#375] add compose varaible to change API BASE URL at runtime
For development, the API base URL is taken from the `.env` file.

```
API_BASE_URL=http://localhost:3001/v1
```

For production, you can use the pre-build docker image which uses the default value set at building time (`http://localhost:3001/v1`).
If you want, you can overwrite it when you run the new app with the env var `NUXT_PUBLIC_API_BASE` but it only works if you are serving the app with node.
It doesn't work for static web app. You new to rebuild the app with your
value for the env var `API_BASE_URL` so that it's included in the
packaged app.